# TinadecCore/tests/__snapshots__/openapi.core.json
{
...
  "info": {
    "title": "TinadecCore.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/sessions/{sessionId}/interactions": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "InteractionsEndpoints"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "InteractionsEndpoints"
        ]
      }
    },
    "/api/v1/sessions/{sessionId}/interactions/{interactionId}/cancel": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "interactionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "InteractionsEndpoints"
        ]
      }
    },
    "/api/v1/sessions/{sessionId}/interactions/{interactionId}/reassign": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "interactionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "InteractionsEndpoints"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
...
}